requirements.txt
flask
qrcode[svg]
pillow
pyzbar
cadquery